+++ test/integration/pack/test_pack_docker.py
@@ -146,25 +146,24 @@ def test_pack(docker_image, tmpdir, docker_client):
         assert installed_version == expected_version
 
 
-def test_invalid_base_runtime_dockerfile(cartridge_cmd, project_without_dependencies, module_tmpdir, tmpdir):
-    invalid_dockerfile_path = os.path.join(tmpdir, 'Dockerfile')
-    with open(invalid_dockerfile_path, 'w') as f:
+def test_custom_base_runtime_dockerfile(cartridge_cmd, project_without_dependencies, module_tmpdir, tmpdir):
+    custom_base_dockerfile_path = os.path.join(tmpdir, 'Dockerfile')
+    with open(custom_base_dockerfile_path, 'w') as f:
         f.write('''
-            # Invalid dockerfile
-            FROM ubuntu:xenial
+            # Non standard base image
+            FROM my-custom-centos-8
         ''')
 
     cmd = [
         cartridge_cmd,
         "pack", "docker",
-        "--from", invalid_dockerfile_path,
+        "--from", custom_base_dockerfile_path,
         project_without_dependencies.path,
     ]
 
     rc, output = run_command_and_get_output(cmd, cwd=module_tmpdir)
-    assert rc == 1
-    assert 'Invalid base runtime Dockerfile' in output
-    assert 'base image must be centos:8' in output
+    assert rc == 0
+    assert 'Image based on centos:8 is expected to be used' in output
 
 
 def test_project_witout_runtime_dockerfile(cartridge_cmd, project_without_dependencies, tmpdir):